requirements.txt
albucore==0.0.24
albumentations==2.0.8
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
brotli==1.2.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
contourpy==1.3.3
cycler==0.12.1
cython==3.2.4
easydict==1.13
fastapi==0.136.0
filelock==3.28.0
flatbuffers==25.12.19
fonttools==4.62.1
fsspec==2026.3.0
gradio==6.12.0
gradio-client==2.4.1
groovy==0.1.2
h11==0.16.0
hf-gradio==0.4.1
hf-xet==1.4.3
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==1.11.0
idna==3.11
imageio==2.37.3
insightface==0.7.3
jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.5.0
lazy-loader==0.5
markdown-it-py==4.0.0
markupsafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
ml-dtypes==0.5.4
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.4
onnx==1.21.0
onnxruntime==1.24.4
opencv-python-headless==4.13.0.92
orjson==3.11.8
packaging==26.1
pandas==3.0.2
pgvector==0.4.2
pillow==12.2.0
prettytable==3.17.0
protobuf==7.34.1
psycopg==3.3.3
psycopg-binary==3.3.3
pydantic==2.13.2
pydantic-core==2.46.2
pydantic-settings==2.13.1
pydub==0.25.1
pygments==2.20.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.26
pytz==2026.1.post1
pyyaml==6.0.3
requests==2.33.1
rich==15.0.0
safehttpx==0.1.7
safetensors==0.7.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.1
semantic-version==2.10.0
setuptools==81.0.0
shellingham==1.5.4
simsimd==6.5.16
six==1.17.0
starlette==1.0.0
stringzilla==4.6.0
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2026.4.11
timm==1.0.26
tomlkit==0.14.0
torch==2.11.0
torchvision==0.26.0
tqdm==4.67.3
typer==0.24.1
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
uvicorn==0.44.0
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
